refactor(test): split the functional suite by concern
functional_test.go had outgrown a single file. Now: the harness (fake
wiring, provider factories, config builders) in functional_harness_test.go,
and one file per behavior area — lifecycles, refusals, approval, stack.
Same ten tests, unchanged, green. The one-file-per-area rule joins
CLAUDE.md so the next area starts its own file.
